Convert Panel component to TSX
